requirements.txt
flask
requests
beautifulsoup4
python-dotenv
anthropic
pytest
pytest-mock
